some work on tcp

+++ cpp_gateway/include/core/basic.hpp
@@ -40,17 +40,17 @@ struct State {
 
 // Motor commands structure
 struct MotorCommands {
-    int16_t m1{0};
-    int16_t m2{0};
-    int16_t m3{0};
-    int16_t m4{0};
+    uint16_t m1{0};
+    uint16_t m2{0};
+    uint16_t m3{0};
+    uint16_t m4{0};
 };
 
 // Actions structure (commands sent to robot)
 struct Actions {
     MotorCommands motors;
     uint16_t beep_ms{0};
-    uint8_t flags{0};
+    uint16_t flags{0};
 };
 
 // Timestamp pair (wall clock and monotonic)
